+++ lib/GenericShell.ts
@@ -1,6 +1,9 @@
+import * as vscode from 'vscode';
 import { ChildProcess, spawn } from 'child_process'
 import { EventEmitter } from 'events'
-import * as utils from './utils';
+import * as fs from 'fs';
+import * as path from 'path';
+import { fstat } from 'fs';
 
 //inspired from https://github.com/WebFreak001/code-debug/blob/master/src/backend/mi2/mi2.ts for inspiration of an EventEmitter 
 const nonOutput = /^(?:\d*|undefined)[\*\+\=]|[\~\@\&\^]/;
@@ -38,6 +41,25 @@ export class GenericShell extends EventEmitter {
         super();
         this.logOutput = logOutput;
         this.shellOutput = shellOutput;
+
+        // Find Erlang 'bin' directory
+        let erlangPath = vscode.workspace.getConfiguration("erlang").get("erlangPath", null);
+        if (erlangPath) {
+            if (erlangPath.match(/^[A-Za-z]:/)) {
+                // Windows absolute path (C:\...) is applicable on Windows only
+                if (process.platform == 'win32') {
+                    this.erlangPath = path.win32.normalize(erlangPath);
+                }
+            } else {
+                erlangPath = path.normalize(erlangPath);
+                if (! fs.existsSync(erlangPath)) {
+                    erlangPath = path.join(vscode.workspace.rootPath, erlangPath);
+                }
+                if (fs.existsSync(erlangPath)) {
+                    this.erlangPath = erlangPath;
+                }
+            }
+        }
     }
 
     protected RunProcess(processName, startDir: string, args: string[]): Promise<number> {