+++ lib/RebarShell.ts
@@ -31,19 +31,17 @@ export default class RebarShell extends GenericShell {
      * @returns Promise resolved or rejected when rebar exits
      */
     public async runScript(cwd: string, commands: string[]): Promise<RebarShellResult> {
-        let args = commands,
-            rebarFileName = this.getRebarFullPath();
-
-        if (process.platform == 'win32') {
-            args = [rebarFileName].concat(args);
-            rebarFileName = 'escript.exe';
-        }
+        // Rebar may not have execution permission (e.g. if extension is built
+        // on Windows but installed on Linux). Let's always run rebar by escript.
+        let escript = (process.platform == 'win32' ? 'escript.exe' : 'escript');
+        let rebarFileName = this.getRebarFullPath();
+        let args = [rebarFileName].concat(commands);
 
         this.shellOutput.clear();
 
         let result: number;
         try {
-            result = await this.RunProcess(rebarFileName, cwd, args);
+            result = await this.RunProcess(escript, cwd, args);
         } catch (exitCode) {
             result = exitCode;
         }