+++ CHANGELOG.md
@@ -14,6 +14,7 @@
     expressions
   * Distinguish function type in type specification and explicit function
     expression
+* Setting `erlang.erlangPath` now can accept workspace relative path
 
 ## Version 0.6.0 (January 23, 2020)